1.1

+++ HomeworkAssignment3/Controllers/borrowsController.cs
@@ -1,55 +1,55 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
 using System.Linq;
-using System.Threading.Tasks;
 using System.Net;
-using System.Web;
+using System.Threading.Tasks;
 using System.Web.Mvc;
 using HomeworkAssignment3.Models; // Ensure this matches your actual namespace
-using HomeworkAssignment3; // If you need other entities
 
 namespace HomeworkAssignment3.Controllers
 {
-    public class borrowsController : Controller
+    public class BorrowsController : Controller
     {
-        private LibraryContext db = new LibraryContext(); // Use LibraryContext instead
+        private readonly LibraryContext db = new LibraryContext(); // Use LibraryContext
 
-        // GET: borrows
+        // GET: Borrows
         public async Task<ActionResult> Index()
         {
             var borrows = db.Borrows.Include(b => b.Book).Include(b => b.Student);
             return View(await borrows.ToListAsync());
         }
 
-        // GET: borrows/Details/5
+        // GET: Borrows/Details/5
         public async Task<ActionResult> Details(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Borrow borrow = await db.Borrows.FindAsync(id);
+
+            var borrow = await db.Borrows.Include(b => b.Book).Include(b => b.Student).FirstOrDefaultAsync(b => b.BorrowId == id);
             if (borrow == null)
             {
                 return HttpNotFound();
             }
+
             return View(borrow);
         }
 
-        // GET: borrows/Create
+        // GET: Borrows/Create
         public ActionResult Create()
         {
             ViewBag.BookId = new SelectList(db.Books, "BookId", "Name");
-            ViewBag.StudentId = new SelectList(db.Students, "StudentId", "Name");
+            ViewBag.StudentId = new SelectList(db.Students, "StudentId", "Name"); // Populate StudentId
             return View();
         }
 
-        // POST: borrows/Create
+
+        // POST: Borrows/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<ActionResult> Create([Bind(Include = "BorrowId,StudentId,BookId,TakenDate,BroughtDate")] Borrow borrow)
+        public async Task<ActionResult> Create([Bind(Include = "StudentId,BookId,TakenDate,BroughtDate")] Borrow borrow)
         {
             if (ModelState.IsValid)
             {
@@ -63,7 +63,7 @@ namespace HomeworkAssignment3.Controllers
             return View(borrow);
         }
 
-        // GET: borrows/Edit/5
+        // GET: Borrows/Edit/5
         public async Task<ActionResult> Edit(int? id)
         {
             if (id == null)
@@ -75,12 +75,14 @@ namespace HomeworkAssignment3.Controllers
             {
                 return HttpNotFound();
             }
+
+            // Populate ViewBag.Students and ViewBag.Books before returning the view
             ViewBag.BookId = new SelectList(db.Books, "BookId", "Name", borrow.BookId);
-            ViewBag.StudentId = new SelectList(db.Students, "StudentId", "Name", borrow.StudentId);
+            ViewBag.StudentId = new SelectList(db.Students, "StudentId", "Name", borrow.StudentId); // Populate StudentId
             return View(borrow);
         }
 
-        // POST: borrows/Edit/5
+        // POST: Borrows/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> Edit([Bind(Include = "BorrowId,StudentId,BookId,TakenDate,BroughtDate")] Borrow borrow)
@@ -91,34 +93,40 @@ namespace HomeworkAssignment3.Controllers
                 await db.SaveChangesAsync();
                 return RedirectToAction("Index");
             }
+
             ViewBag.BookId = new SelectList(db.Books, "BookId", "Name", borrow.BookId);
             ViewBag.StudentId = new SelectList(db.Students, "StudentId", "Name", borrow.StudentId);
             return View(borrow);
         }
 
-        // GET: borrows/Delete/5
+        // GET: Borrows/Delete/5
         public async Task<ActionResult> Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Borrow borrow = await db.Borrows.FindAsync(id);
+
+            var borrow = await db.Borrows.Include(b => b.Book).Include(b => b.Student).FirstOrDefaultAsync(b => b.BorrowId == id);
             if (borrow == null)
             {
                 return HttpNotFound();
             }
+
             return View(borrow);
         }
 
-        // POST: borrows/Delete/5
+        // POST: Borrows/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(int id)
         {
-            Borrow borrow = await db.Borrows.FindAsync(id);
-            db.Borrows.Remove(borrow);
-            await db.SaveChangesAsync();
+            var borrow = await db.Borrows.FindAsync(id);
+            if (borrow != null)
+            {
+                db.Borrows.Remove(borrow);
+                await db.SaveChangesAsync();
+            }
             return RedirectToAction("Index");
         }
